requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.14.0
dnspython==2.6.1
psycopg==3.1.18
psycopg2==2.9.9
pymongo==4.6.2
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
